# API/openapi (3).json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyTree Banking API",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://footballfriendlies-582a7.nw.r.appspot.com"
    }
  ],
...
  "paths": {
    "/api/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "returns all companies stored within the database",
        "description": "Gets all company information from all companies in the database",
        "operationId": "GetAllCompanies",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTransaction"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
    "/api/companies/similarCompanies/{accountNumber}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Finds similar companies with better RAG score",
        "description": "Place account number in URL after similarCompanies",
        "operationId": "FindSimilarCompanies",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "path",
            "description": "Account number of company.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid account number"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/api/companies/updateEnvironmentalImpactScore": {
      "put": {
        "tags": [
          "companies"
        ],
        "summary": "Update a specific companies environmental impact score",
        "description": "takes the company ID and all the scores, or just single ones in the body.",
        "operationId": "UpdateEnvironmentalImpactScore",
        "parameters": [
          {
            "name": "Account Number",
            "in": "query",
            "description": "Account number of company which you want to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "xxxxxxxxx"
            }
          },
          {
            "name": "Sustainability Pratices",
            "in": "query",
            "description": "Score from 0 to 10 representing their sustainabillity practices",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Waste Management",
            "in": "query",
            "description": "Score from 0 to 10 representing their waste management practices",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Carbon Emissions",
            "in": "query",
            "description": "Score from 0 to 10 representing their carbon emissions",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Company not Found"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/api/companies/lazy_load_company": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Gets just the company names and corresponding account numbers from the database",
        "description": "",
        "operationId": "lazyLoadCompany",
        "responses": {
          "200": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LazyLoadCompany"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/user_transactions": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Gets a users transactions",
        "description": "",
        "operationId": "GetUserTransactions",
        "parameters": [
          {
            "name": "accountNumber",
            "in": "query",
            "description": "Account number of user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "User Not Found"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/api/user_transactions/to": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Returns user transactions to a specific company",
        "description": "Returns user transactions to a specific company defined by the company account number",
        "operationId": "GetUserTransactionTo",
        "parameters": [
          {
            "name": "CompanyAccountNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          {
            "name": "UserAccountNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTransactionToRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTransactionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/user/update_balance": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Update user balance",
        "description": "Increment/decrement user balance",
        "operationId": "UpdateUserBalance",
        "parameters": [
          {
            "name": "UserAccountNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "BalanceDifference",
            "in": "query",
            "description": "+- value",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateBalanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateBalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "422": {
            "description": "Not enough user funds."
          }
        }
      }
    },
    "/api/user/add": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Add new User",
        "description": "",
        "operationId": "AddNewUser",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "First name of new user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age",
            "in": "query",
            "required": true,
            "description": "Age of the new user",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "accountbalance",
            "in": "query",
            "required": true,
            "description": "starting balance of the new user",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserTransactionRequest": {
        "type": "object",
        "properties": {
          "UserAccountNumber": {
            "type": "integer",
            "format": "int64",
            "example": 12345678910
          }
        },
        "xml": {
          "name": "UserTransaction"
        }
      },
      "UserTransactionResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66ed514ba8eac352649e9f0a"
          },
          "from": {
            "type": "integer",
            "format": "int64",
            "example": 12345678910
          },
          "to": {
            "type": "string",
            "example": "123456789"
          },
          "Time": {
            "type": "string",
            "example": "2024-09-05T05:12:46.030Z"
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "example": 200
          }
        }
      },
      "UserTransactionToRequest": {
        "type": "object",
        "properties": {
          "CompanyAccountNumber": {
            "type": "string",
            "example": "123456789"
          },
          "UserAccountNumber": {
            "type": "integer",
            "format": "int64",
            "example": 12345678910
          }
        },
        "xml": {
          "name": "UserTransactionTo"
        }
      },
...
      "LazyLoadCompany": {
        "type": "object",
        "properties": {
          "CompanyName": {
            "type": "string",
            "example": "RedFuel Services"
          },
          "AccountNumber": {
            "type": "string",
            "example": "000000001"
          }
        }
      },
...
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}